+++ scripts/dzVents/runtime/tests/scripts/script4.lua
@@ -1,7 +1,7 @@
 return {
 	active = true,
 	on = {
-		devices = {'onscript4'}
+		devices = {'onscript4', 8}
 	},
 	execute = function(domoticz, device)
 		return 'script4'